feat: :sparkles: Add POST endpoint and express-validator validations
New feature where a POST endpoint for creating a doctor is added. It also includes data validations using express-validator to ensure that the data received in the request is in the correct format and meets the necessary criteria. This will help maintain data integrity and provide better error messages to the client.